Deploy2

+++ models/author.go
@@ -19,9 +19,9 @@ type Author struct {
 }
 
 func GetAllAuthors(authToken string) ([]Author, error) {
-	// if !ValidToken(authToken) {
-	// 	return []Author{}, errors.New("Unauthorized")
-	// }
+	if !ValidToken(authToken) {
+		return []Author{}, errors.New("Unauthorized")
+	}
 	var authors []Author
 	var (
 		id         int
@@ -64,9 +64,9 @@ func GetAllAuthors(authToken string) ([]Author, error) {
 }
 
 func FindAuthor(authorId int, authToken string) (interface{}, error) {
-	// if !ValidToken(authToken) {
-	// 	return []Author{}, errors.New("Unauthorized")
-	// }
+	if !ValidToken(authToken) {
+		return []Author{}, errors.New("Unauthorized")
+	}
 	var (
 		id         int
 		first_name string
@@ -97,10 +97,10 @@ func FindAuthor(authorId int, authToken string) (interface{}, error) {
 	return author, nil
 }
 
-func CreateAuthor(body io.Reader, authToken string) (GeneralResponse, error) {
-	// if !ValidToken(authToken) {
-	// 	return []Author{}, errors.New("Unauthorized")
-	// }
+func CreateAuthor(body io.Reader, authToken string) (interface{}, error) {
+	if !ValidToken(authToken) {
+		return []Author{}, errors.New("Unauthorized")
+	}
 	var a Author
 	err := json.NewDecoder(body).Decode(&a)
 	if err != nil {
@@ -120,10 +120,10 @@ func CreateAuthor(body io.Reader, authToken string) (GeneralResponse, error) {
 	}
 }
 
-func UpdateAuthor(AuthorId int, body io.Reader, authToken string) (GeneralResponse, error) {
-	// if !ValidToken(authToken) {
-	// 	return []Author{}, errors.New("Unauthorized")
-	// }
+func UpdateAuthor(AuthorId int, body io.Reader, authToken string) (interface{}, error) {
+	if !ValidToken(authToken) {
+		return []Author{}, errors.New("Unauthorized")
+	}
 	var a Author
 	err := json.NewDecoder(body).Decode(&a)
 	query := `
@@ -146,9 +146,9 @@ func UpdateAuthor(AuthorId int, body io.Reader, authToken string) (GeneralRespon
 }
 
 func DeleteAuthor(authorId int, authToken string) (GeneralResponse, error) {
-	// if !ValidToken(authToken) {
-	// 	return GeneralResponse{Message: "Unauthorized"}, errors.New("Unauthorized")
-	// }
+	if !ValidToken(authToken) {
+		return GeneralResponse{Message: "Unauthorized"}, errors.New("Unauthorized")
+	}
 	query := `DELETE FROM authors WHERE id=$1`
 	res, err := database.DB.Exec(query, authorId)
 	rowCount, err := res.RowsAffected()
@@ -162,49 +162,52 @@ func DeleteAuthor(authorId int, authToken string) (GeneralResponse, error) {
 	return GeneralResponse{Message: "Author deleted successfully"}, nil
 }
 
-func GetAuthorsForPaper(paperId int, kawaiiChan chan []Author) {
-	var authors []Author
-	var (
-		id         int
-		first_name string
-		last_name  string
-		created_at time.Time
-		updated_at time.Time
-	)
-	query := `
-		SELECT authors.* FROM authors 
-		INNER JOIN author_papers ON authors.id = author_papers.author_id
-		INNER JOIN papers ON papers.id = author_papers.paper_id
-		WHERE papers.id = $1
-	`
-	rows, err := database.DB.Query(query, paperId)
-	if err != nil {
-		fmt.Println(err)
-	}
-	defer rows.Close()
-	for rows.Next() {
-		err = rows.Scan(
-			&id,
-			&first_name,
-			&last_name,
-			&created_at,
-			&updated_at,
+func GetAuthorsForPaper(paperId int) <-chan []Author {
+	ch := make(chan []Author)
+	go func() {
+		var authors []Author
+		var (
+			id         int
+			first_name string
+			last_name  string
+			created_at time.Time
+			updated_at time.Time
 		)
+		query := `
+			SELECT authors.* FROM authors 
+			INNER JOIN author_papers ON authors.id = author_papers.author_id
+			INNER JOIN papers ON papers.id = author_papers.paper_id
+			WHERE papers.id = $1
+		`
+		rows, err := database.DB.Query(query, paperId)
 		if err != nil {
 			fmt.Println(err)
 		}
-		author := Author{
-			Id:        id,
-			FirstName: first_name,
-			LastName:  last_name,
-			CreatedAt: created_at.String(),
-			UpdatedAt: updated_at.String(),
+		defer rows.Close()
+		for rows.Next() {
+			err = rows.Scan(
+				&id,
+				&first_name,
+				&last_name,
+				&created_at,
+				&updated_at,
+			)
+			if err != nil {
+				fmt.Println(err)
+			}
+			author := Author{
+				Id:        id,
+				FirstName: first_name,
+				LastName:  last_name,
+				CreatedAt: created_at.String(),
+				UpdatedAt: updated_at.String(),
+			}
+			authors = append(authors, author)
 		}
-		authors = append(authors, author)
-	}
-
-	if err != nil {
-		fmt.Println("Error getting paper's authors", err)
-	}
-	kawaiiChan <- authors
+		if err != nil {
+			fmt.Println("Error getting paper's authors", err)
+		}
+		ch <- authors
+	}()
+	return ch
 }
