+++ models/figure.go
@@ -20,9 +20,9 @@ type Figure struct {
 }
 
 func GetAllFigures(authToken string) ([]Figure, error) {
-	// if !ValidToken(authToken) {
-	// 	return []Figure{}, errors.New("Unauthorized")
-	// }
+	if !ValidToken(authToken) {
+		return []Figure{}, errors.New("Unauthorized")
+	}
 	var figures []Figure
 	var (
 		id          int
@@ -68,9 +68,9 @@ func GetAllFigures(authToken string) ([]Figure, error) {
 }
 
 func FindFigure(figureId int, authToken string) (interface{}, error) {
-	// if !ValidToken(authToken) {
-	// 	return []Figure{}, errors.New("Unauthorized")
-	// }
+	if !ValidToken(authToken) {
+		return []Figure{}, errors.New("Unauthorized")
+	}
 	var (
 		id          int
 		name        string
@@ -105,9 +105,9 @@ func FindFigure(figureId int, authToken string) (interface{}, error) {
 }
 
 func CreateFigure(body io.Reader, authToken string) (interface{}, error) {
-	// if !ValidToken(authToken) {
-	// 	return []Figure{}, errors.New("Unauthorized")
-	// }
+	if !ValidToken(authToken) {
+		return []Figure{}, errors.New("Unauthorized")
+	}
 	var f Figure
 	err := json.NewDecoder(body).Decode(&f)
 	if err != nil {
@@ -127,10 +127,10 @@ func CreateFigure(body io.Reader, authToken string) (interface{}, error) {
 	}
 }
 
-func UpdateFigure(figureId int, body io.Reader, authToken string) (GeneralResponse, error) {
-	// if !ValidToken(authToken) {
-	// 	return []Figure{}, errors.New("Unauthorized")
-	// }
+func UpdateFigure(figureId int, body io.Reader, authToken string) (interface{}, error) {
+	if !ValidToken(authToken) {
+		return []Figure{}, errors.New("Unauthorized")
+	}
 	var f Figure
 	err := json.NewDecoder(body).Decode(&f)
 	query := `
@@ -154,9 +154,9 @@ func UpdateFigure(figureId int, body io.Reader, authToken string) (GeneralRespon
 }
 
 func DeleteFigure(figureId int, authToken string) (GeneralResponse, error) {
-	// if !ValidToken(authToken) {
-	// 	return GeneralResponse{Message: "Unauthorized"}, errors.New("Unauthorized")
-	// }
+	if !ValidToken(authToken) {
+		return GeneralResponse{Message: "Unauthorized"}, errors.New("Unauthorized")
+	}
 	query := `DELETE FROM figures WHERE id=$1`
 	res, err := database.DB.Exec(query, figureId)
 	rowCount, err := res.RowsAffected()
@@ -170,52 +170,55 @@ func DeleteFigure(figureId int, authToken string) (GeneralResponse, error) {
 	return GeneralResponse{Message: "Figure deleted successfully"}, nil
 }
 
-func GetFiguresForPaper(paperId int, kawaiiChan chan []Figure) {
-	var figures []Figure
-	var (
-		id          int
-		name        string
-		figure_type string
-		image_file  string
-		created_at  time.Time
-		updated_at  time.Time
-	)
-	query := `
-		SELECT figures.* FROM figures 
-		INNER JOIN figure_papers ON figures.id = figure_papers.figure_id
-		INNER JOIN papers ON papers.id = figure_papers.paper_id
-		WHERE papers.id = $1
-	`
-	rows, err := database.DB.Query(query, paperId)
-	if err != nil {
-		fmt.Println(err)
-	}
-	defer rows.Close()
-	for rows.Next() {
-		err = rows.Scan(
-			&id,
-			&name,
-			&figure_type,
-			&image_file,
-			&created_at,
-			&updated_at,
+func GetFiguresForPaper(paperId int) <-chan []Figure {
+	ch := make(chan []Figure)
+	go func() {
+		var figures []Figure
+		var (
+			id          int
+			name        string
+			figure_type string
+			image_file  string
+			created_at  time.Time
+			updated_at  time.Time
 		)
+		query := `
+			SELECT figures.* FROM figures 
+			INNER JOIN figure_papers ON figures.id = figure_papers.figure_id
+			INNER JOIN papers ON papers.id = figure_papers.paper_id
+			WHERE papers.id = $1
+		`
+		rows, err := database.DB.Query(query, paperId)
 		if err != nil {
 			fmt.Println(err)
 		}
-		figure := Figure{
-			Id:         id,
-			Name:       name,
-			FigureType: figure_type,
-			ImageFile:  image_file,
-			CreatedAt:  created_at.String(),
-			UpdatedAt:  updated_at.String(),
+		defer rows.Close()
+		for rows.Next() {
+			err = rows.Scan(
+				&id,
+				&name,
+				&figure_type,
+				&image_file,
+				&created_at,
+				&updated_at,
+			)
+			if err != nil {
+				fmt.Println(err)
+			}
+			figure := Figure{
+				Id:         id,
+				Name:       name,
+				FigureType: figure_type,
+				ImageFile:  image_file,
+				CreatedAt:  created_at.String(),
+				UpdatedAt:  updated_at.String(),
+			}
+			figures = append(figures, figure)
 		}
-		figures = append(figures, figure)
-	}
-
-	if err != nil {
-		fmt.Println("Error getting paper's figures", err)
-	}
-	kawaiiChan <- figures
+		if err != nil {
+			fmt.Println("Error getting paper's figures", err)
+		}
+		ch <- figures
+	}()
+	return ch
 }
